requirements.txt
django
django_rest_framework
psycopg2-binary
Pillow
mysqlclient

